Track ping and channel statistics only in development environments
In production a lot of logs can cause heavy lags on the client side.
Also, the ping only really works when running both instances on
the same device as the system clocks need to be in perfect sync.

Ultimately, it would be nice to detect if the Lobby is in dev-mode
and only write timestamps to packets if that is the case.
I will look into that later, but for now I hope that
ignoring the timestamps if not in dev-mode is sufficient.